Reorganize app, add refresh link
