+++ README.md
@@ -2,7 +2,10 @@
 
 This is an API and matching React Native app to allow users to easily see the Greene County Ohio Alert Center.
 
-`/api` for the API
-`/app` for the React Native App
+![Screenshot of the first iteration of the mobile app without any alerts active](/screenshot_no_alerts.png)
 
-![Screenshot of mobile app without any alerts active](/screenshot_no_alerts.png)
+## About this repo
+
+- `/api` for the API
+
+- `/app` for the React Native App
