+++ app/src/ApolloWrapper.js
@@ -1,0 +1,16 @@
+import React from "react";
+import { ApolloProvider } from "react-apollo";
+import { ApolloClient } from "apollo-client";
+import { HttpLink } from "apollo-link-http";
+import { InMemoryCache } from "apollo-cache-inmemory";
+
+const client = new ApolloClient({
+  link: new HttpLink({ uri: "https://greene-co-alerts.now.sh" }),
+  cache: new InMemoryCache()
+});
+
+const ApolloWrapper = ({ children }) => (
+  <ApolloProvider client={client}>{children}</ApolloProvider>
+);
+
+export default ApolloWrapper;
