+++ app/src/theme.js
@@ -1,0 +1,1 @@
+export const appColor = "#e67e22";
